+++ lizard_damage/management/commands/damage.py
@@ -29,8 +29,8 @@ class Command(BaseCommand):
             settings.DATA_ROOT, 'waterlevel', 'ws%i.asc' % i
         ) for i in range(5)]
         ds_wl_filenames = [
-            os.path.join(settings.DATA_ROOT, 'waterlevel', 'utrechtq.asc'),
-            #os.path.join(settings.DATA_ROOT, 'waterlevel', 'i37_en2_09_delfgauw.asc'),
+            #os.path.join(settings.DATA_ROOT, 'waterlevel', 'utrechtq.asc'),
+            os.path.join(settings.DATA_ROOT, 'waterlevel', 'i37_en2_09_delfgauw.asc'),
             #os.path.join(settings.DATA_ROOT, 'waterlevel', 'goudswaard.asc'),
             #os.path.join(settings.DATA_ROOT, 'waterlevel', '1_ha_gras.asc'),
             #os.path.join(settings.DATA_ROOT, 'waterlevel', 'ws1.asc'),