+++ CHANGES.rst
@@ -2,11 +2,20 @@ Changelog of lizard-damage
 ===================================================
 
 
-0.6.8 (unreleased)
+0.6.9 (unreleased)
 ------------------
 
 - Zip results after each tile -> else the harddisk will be flooded when 
   calculating big scenarios.
+
+
+0.6.8 (2012-10-02)
+------------------
+
+- Replace migrations by one initial migration, that excludes the models
+  AhnIndex and Roads
+
+- Make AhnIndex refer to raster server via router.
 
 
 0.6.7 (2012-10-01)